requirements.txt
transformers
torch
datasets
soundfile
accelerate
bitsandbytes
neo4j
sentence-transformers
pyaudio
numpy
scipy
flask
flask-cors
requests
google-generativeai
langchain
langchain-google-genai
pydub
webrtcvad
python-dotenv
streamlit
plotly
pandas
nltk
spacy
networkx
matplotlib
seaborn
cryptography